+++ tests/testRecons64bit.cc
@@ -92,7 +92,7 @@ testRecons4DirPlus::run(void)
      for(int spin=0; spin < 4; spin++) { 
        for(int reim=0; reim < 2; reim++) { 
 	 double diff = spinor1[spin][col][reim] - spinor2[spin][col][reim];
-	 diff /= 4*3*2;
+	 diff /= (double)(4*3*2);
 	 
 #if 0
 	 QDPIO::cout << "  col=" << col 
@@ -190,7 +190,7 @@ testRecons4DirMinus::run(void)
      for(int spin=0; spin < 4; spin++) { 
        for(int reim=0; reim < 2; reim++) { 
 	 double diff = spinor1[spin][col][reim] - spinor2[spin][col][reim];
-	 diff /= 4*3*2;
+	 diff /= (double)(4*3*2);
 #if 0
 	 QDPIO::cout << "  col=" << col 
 		     << " sp=" << spin
